+++ src/stim300_constants.h
@@ -3,6 +3,7 @@
 
 #include <cstdint>
 #include <math.h>
+#include <array>
 
 namespace stim_300
 {
@@ -129,7 +130,7 @@ constexpr DatagramIdentifier rawToDatagramIdentifier(uint8_t datagram_id)
     case 0xBD:
       return DatagramIdentifier::CONFIGURATION_CRLF;
     default:
-      throw std::out_of_range("Raw value does not match any implemented datgram ids");
+      return DatagramIdentifier::CONFIGURATION;  // TODO fix error handeling
   }
 };
 
@@ -174,6 +175,81 @@ constexpr uint8_t numberOfPaddingBytes(DatagramIdentifier datagram_identifier)
     case DatagramIdentifier::RATE_ACC_INCL_TEMP_AUX:
       return 1;
   }
+}
+
+enum SensorIndx
+{
+  GYRO = 0,
+  ACC,
+  INCL,
+  TEMP,
+  AUX
+};
+
+constexpr std::array<bool, 5> isIncluded(DatagramIdentifier datagram_identifier)
+{
+  switch (datagram_identifier)
+  {
+    // rate,  acc, incl, temp,  aux
+    case DatagramIdentifier::RATE:
+      return { true, false, false, false, false };
+    case DatagramIdentifier::RATE_ACC:
+      return { true, true, false, false, false };
+    case DatagramIdentifier::RATE_INCL:
+      return { true, false, true, false, false };
+    case DatagramIdentifier::RATE_ACC_INCL:
+      return { true, true, true, false, false };
+    case DatagramIdentifier::RATE_TEMP:
+      return { true, false, false, true, false };
+    case DatagramIdentifier::RATE_ACC_TEMP:
+      return { true, true, false, true, false };
+    case DatagramIdentifier::RATE_INCL_TEMP:
+      return { true, false, true, true, false };
+    case DatagramIdentifier::RATE_ACC_INCL_TEMP:
+      return { true, true, true, true, false };
+    case DatagramIdentifier::RATE_AUX:
+      return { true, false, false, false, true };
+    case DatagramIdentifier::RATE_ACC_AUX:
+      return { true, true, false, false, true };
+    case DatagramIdentifier::RATE_INCL_AUX:
+      return { true, false, true, false, true };
+    case DatagramIdentifier::RATE_ACC_INCL_AUX:
+      return { true, true, true, false, true };
+    case DatagramIdentifier::RATE_TEMP_AUX:
+      return { true, false, false, true, true };
+    case DatagramIdentifier::RATE_ACC_TEMP_AUX:
+      return { true, true, false, true, true };
+    case DatagramIdentifier::RATE_INCL_TEMP_AUX:
+      return { true, false, true, true, true };
+    case DatagramIdentifier::RATE_ACC_INCL_TEMP_AUX:
+      return { true, true, true, true, true };
+    case DatagramIdentifier::CONFIGURATION:
+      return { false, false, false, false, false };
+    case DatagramIdentifier::CONFIGURATION_CRLF:
+      return { false, false, false, false, false };
+    default:
+      return { false, false, false, false, false };
+      // throw std::out_of_range("Undefined datagram identifier");
+  }
+}
+
+static const uint8_t calculateDatagramSize(DatagramIdentifier datagram_identifier)
+{
+  std::array<bool, 5> is_included = isIncluded(datagram_identifier);
+  uint8_t n_inertial_sensors{ 1 };
+  n_inertial_sensors += is_included[SensorIndx::ACC] ? 1 : 0;
+  n_inertial_sensors += is_included[SensorIndx::INCL] ? 1 : 0;
+
+  uint8_t size{ 0 };
+  size += N_BYTES_DATAGRAM_ID;
+  size += n_inertial_sensors * (3 * N_BYTES_INERTIAL_SENSOR + N_BYTES_STATUS);
+  size += is_included[SensorIndx::TEMP] ? n_inertial_sensors * (3 * N_BYTES_TEMP_SENSOR + N_BYTES_STATUS) : 0;
+  size += is_included[SensorIndx::AUX] ? N_BYTES_AUX_SENSOR + N_BYTES_STATUS : 0;
+  size += N_BYTES_COUNTER;
+  size += N_BYTES_LATENCY;
+  size += N_BYTES_CRC;
+
+  return size;
 }
 
 }  // namespace stim_300
